fast_gate/requirements.txt
aio-pika==9.4.1
aiormq==6.8.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
black==24.3.0
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.110.1
grpcio==1.62.2
grpcio-tools==1.62.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.0
pamqp==3.3.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.5.0
protobuf==4.25.3
pyasn1==0.6.0
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
pytest==8.2.0
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.1
redis==5.0.4
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
